refactor: move share dialog to a new component and make the share button reutilize the share dialog
